[Keyboard] Fix 9key macropad keymap
Fails to compile on the configurator because tapdance is enabled for the entire keyboard
