+++ keyboards/9key/keymaps/default/keymap.c
@@ -1,16 +1,5 @@
 #include QMK_KEYBOARD_H
 
-// Tap Dance Declarations
-enum {
-  ENT_5 = 0,
-  ZERO_7
-};
-
-// Macro Declarations
-enum {
-  DBL_0 = 0
-};
-
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
 
 /* LAYER 0
@@ -24,8 +13,8 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
  */
 [0] = LAYOUT( \
   KC_1,       KC_2,      KC_3,       \
-  KC_4,       TD(ENT_5), KC_6,       \
-  TD(ZERO_7), KC_8,      LT(1, KC_9) \
+  KC_4,       KC_5,      KC_6,       \
+  KC_7,       KC_8,      LT(1, KC_9) \
 ),
 
 /* LAYER 1
@@ -39,27 +28,8 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
  */
 [1] = LAYOUT( \
   KC_ESC,   KC_PLUS, KC_MINS, \
-  KC_BSPC,  KC_ASTR, KC_SLSH, \
-  M(DBL_0), KC_DOT,  KC_TRNS  \
+  KC_ENTER, KC_ASTR, KC_SLSH, \
+  KC_0,     KC_DOT,  KC_TRNS  \
 )
 
 };
-
-qk_tap_dance_action_t tap_dance_actions[] = {
-  [ENT_5] = ACTION_TAP_DANCE_DOUBLE(KC_5, KC_ENT),
-  [ZERO_7] = ACTION_TAP_DANCE_DOUBLE(KC_7, KC_0)
-};
-
-const macro_t *action_get_macro(keyrecord_t *record, uint8_t id, uint8_t opt) {
-  if (record->event.pressed) {
-      switch(id) {
-          case DBL_0:
-              SEND_STRING("00");
-              return false;
-      }
-  }
-  return MACRO_NONE;
-};
-
-void matrix_init_user(void) {
-}
