+++ keyboards/9key/keymaps/tap_dance/rules.mk
@@ -1,0 +1,1 @@
+TAP_DANCE_ENABLE = yes
